+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add **float perimeter_Parker_fast()** for completeness.
 - add **float perimeter_polynome2()** quadratic polynome approximation.
 - add **float getH()** extra attribute.
+- add **float ratio()** extra attribute.
 - update GitHub actions
 - update / add examples
 - update readme.md (a.o performance table)
+- update unit tests
 - minor edits
 
 ----